requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
cryptography==42.0.5
passlib==1.7.4
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
weasyprint==61.0
jinja2==3.1.3
jinja2-time==0.2.0
apscheduler==3.10.4
aiogram==3.4.1
redis==5.0.3
